+++ src/components/Alert/styles.scss
