+++ src/components/HealthInformation/styles.scss
