+++ src/components/Navbar/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NavLink, useLocation } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import Logo from '../Logo/Logo';
+import Logo from '../Logo';
 import Facebook from '../../images/facebook.png';
 import Twitter from '../../images/twitter.png';
