+++ src/containers/Doctors/index.js
@@ -6,9 +6,9 @@ import {
   getDoctors,
   getDoctorsBySpeciality,
 } from '../../store/actions/doctors';
-import './Doctors.scss';
-import Doctor from '../../components/Doctor/Doctor';
-import Loading from '../../components/Loading/Loading';
+import './styles.scss';
+import Doctor from '../../components/Doctor';
+import Loading from '../../components/Loading';
 
 const Doctors = ({
   doctors, getDoctors, isLoading, getDoctorsBySpeciality, speciality,
