+++ src/containers/Register/index.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { Link, Redirect } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { registerUser } from '../../store/actions/auth';
-import './Register.scss';
+import './styles.scss';
 
 const Register = ({ isLoggedIn, registerUser, isLoading }) => {
   const [name, setName] = useState('');
