+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import App from './containers/App/App';
+import App from './containers/App';
 import * as serviceWorker from './serviceWorker';
 import './index.css';
 import store from './store';
